# maestro-hive/sunday_com_backup_20251005/backend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Sunday.com API
  description: |
    Comprehensive API documentation for Sunday.com work management platform.

    ## Authentication
    All API endpoints require authentication via JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting
    - 1000 requests per hour for authenticated users
    - 100 requests per hour for unauthenticated users

    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "error": {
        "type": "validation_error",
        "message": "Invalid request data",
        "details": {},
        "requestId": "uuid"
      }
    }
    ```

    ## Pagination
    List endpoints support cursor and offset-based pagination:
    ```
    GET /api/v1/boards?page=1&limit=20
    ```

  version: 1.0.0
  contact:
    name: Sunday.com Team
    email: api@sunday.com
    url: https://sunday.com/docs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sunday.com/terms
servers:
  - url: https://api.sunday.com/api/v1
    description: Production server
  - url: https://staging-api.sunday.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - bearerAuth: []

# ============================================================================
# AUTHENTICATION
# ============================================================================
paths:
...
  /items:
    post:
      tags: [Items]
      summary: Create item
      description: Create a new item on a board
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ItemWithRelations'
        '422':
          $ref: '#/components/responses/ValidationError'
  /items/{itemId}:
    get:
      tags: [Items]
      summary: Get item
      description: Get item details with related data
      parameters:
        - $ref: '#/components/parameters/ItemId'
        - in: query
          name: includeComments
          schema:
            type: boolean
            default: false
          description: Include item comments
        - in: query
          name: includeChildren
          schema:
            type: boolean
            default: false
          description: Include child items
      responses:
        '200':
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ItemWithRelations'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Items]
      summary: Update item
      description: Update item details and properties
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/ItemWithRelations'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Items]
      summary: Delete item
      description: Soft delete an item
      parameters:
        - $ref: '#/components/parameters/ItemId'
      responses:
        '204':
          description: Item deleted successfully
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /items/{itemId}/move:
    post:
      tags: [Items]
      summary: Move item
      description: Move an item to a different position or board
      parameters:
        - $ref: '#/components/parameters/ItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveItemRequest'
      responses:
        '200':
          description: Item moved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Item'
        '422':
          $ref: '#/components/responses/ValidationError'

# ============================================================================
# COMPONENTS
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Item ID
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          format: uri
          nullable: true
        timezone:
          type: string
          default: "UTC"
        locale:
          type: string
          default: "en"
        settings:
          type: object
          additionalProperties: true
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
        description:
          type: string
          nullable: true
        itemData:
          type: object
          additionalProperties: true
          description: Custom field data
        position:
          type: number
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ItemWithRelations:
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          properties:
            assignments:
              type: array
              items:
                $ref: '#/components/schemas/ItemAssignment'
            parent:
              $ref: '#/components/schemas/Item'
            children:
              type: array
              items:
                $ref: '#/components/schemas/Item'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
            creator:
              $ref: '#/components/schemas/User'
            _count:
              type: object
              properties:
                comments:
                  type: integer
                children:
                  type: integer
                dependencies:
                  type: integer
    ItemAssignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        assignedBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        itemId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
...
    CreateItemRequest:
      type: object
      required: [name, boardId]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        boardId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        itemData:
          type: object
          additionalProperties: true
          description: Custom field data (status, priority, dates, etc.)
        position:
          type: number
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        itemData:
          type: object
          additionalProperties: true
        position:
          type: number
        assigneeIds:
          type: array
          items:
            type: string
            format: uuid
...
    MoveItemRequest:
      type: object
      properties:
        boardId:
          type: string
          format: uuid
          description: Target board ID (for moving to different board)
        parentId:
          type: string
          format: uuid
          description: New parent item ID
        position:
          type: number
          description: New position value

    # Error Schemas
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [type, message]
          properties:
            type:
              type: string
              example: "validation_error"
            message:
              type: string
              example: "Invalid request data"
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "not_found_error"
              message: "Resource not found"
              requestId: "req_123456"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              type: "validation_error"
              message: "Invalid request data"
              details:
                name: ["Name is required"]
                email: ["Invalid email format"]
              requestId: "req_123456"
...